Refractor front end code.
