+++ client/react/components/guided/guidedHeader.js
