+++ client/react/components/guided/guidedPlugins.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Link } from 'react-router';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { backend, testing } from '../actions/actionhelper';
-import { taskRunner, changeDisplayType, changeCheckoutFormat } from '../actions/index';
+import { taskRunner, taskCreation } from '../../actions/actionhelper';
+import { addTask, changeDisplayType, changeGuidedPage } from '../../actions/index';
 
 import { Grid, Row, Col, Button, Jumbotron, PageHeader, Image } from 'react-bootstrap';
 
-class GuidedTaskRunner extends React.Component {
+class GuidedPlugins extends React.Component {
 
   render() {
     return (
@@ -19,8 +19,8 @@ class GuidedTaskRunner extends React.Component {
                bsStyle='primary'
                onClick={
                 ()=> {
-                  this.props.taskRunner('Gulp');
-                  this.props.changeCheckoutFormat(testing);
+                  // this.props.taskRunner('Gulp');
+                  this.props.changeGuidedPage(taskCreation);
                   }
                 }
                block> Gulp </Button>
@@ -30,8 +30,8 @@ class GuidedTaskRunner extends React.Component {
                 bsStyle='danger'
                 onClick={
                 ()=> {
-                  this.props.taskRunner('Grunt');
-                  this.props.changeCheckoutFormat(testing);
+                  // this.props.taskRunner('Grunt');
+                  this.props.changeGuidedPage(taskCreation);
                   }
                 }
                 block> Grunt </Button>
@@ -41,8 +41,8 @@ class GuidedTaskRunner extends React.Component {
                 bsStyle='warning'
                 onClick={
                 ()=> {
-                  this.props.taskRunner('Gulp');
-                  this.props.changeCheckoutFormat(testing);
+                  // this.props.taskRunner('Gulp');
+                  this.props.changeGuidedPage(taskCreation);
                   }
                 }
                 block> Gulp </Button>
@@ -53,7 +53,7 @@ class GuidedTaskRunner extends React.Component {
               <Button
               	bsSize='large'
               	className='navButtons'
-              	onClick={()=> this.props.changeCheckoutFormat(backend)}
+              	onClick={()=> this.props.changeGuidedPage(taskRunner)}
               	block>
               Back
               </Button>
@@ -61,7 +61,7 @@ class GuidedTaskRunner extends React.Component {
             <Col xs={3} xsOffset={1}>
               <Button
               	bsSize='large'
-              	onClick={()=> this.props.changeCheckoutFormat(testing)}
+              	onClick={()=> this.props.changeGuidedPage(taskCreation)}
               	block>
                 Next
               </Button>
@@ -82,10 +82,10 @@ function mapStateToProps(state) {
 
 function matchDispatchToProps(dispatch) {
   return bindActionCreators({
-    taskRunner: taskRunner,
+    addTask: addTask,
     changeDisplayType: changeDisplayType,
-    changeCheckoutFormat: changeCheckoutFormat
+    changeGuidedPage: changeGuidedPage
   }, dispatch);
 }
 
-export default connect(mapStateToProps, matchDispatchToProps)(GuidedTaskRunner);
+export default connect(mapStateToProps, matchDispatchToProps)(GuidedPlugins);
