--- server/bundles/bundle.txt
@@ -1,1 +1,0 @@
-bundle.txt
